requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.1
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pdfminer.six==20200726
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
